+++ include/utils/cifar100_data_loader.hpp
@@ -207,7 +207,7 @@ public:
    * Get a specific batch size (supports both pre-computed and on-demand
    * batches)
    */
-  bool get_batch(int batch_size, Tensor<T> &batch_data,
+  bool get_batch(size_t batch_size, Tensor<T> &batch_data,
                  Tensor<T> &batch_labels) override {
 
     if (batches_prepared_ && batch_size == this->batch_size_) {
@@ -218,31 +218,28 @@ public:
       return false;
     }
 
-    const int actual_batch_size = std::min(
-        batch_size, static_cast<int>(data_.size() - this->current_index_));
+    const size_t actual_batch_size = std::min(
+        batch_size, data_.size() - this->current_index_);
     const int num_classes = use_coarse_labels_
                                 ? cifar100_constants::NUM_COARSE_CLASSES
                                 : cifar100_constants::NUM_CLASSES;
 
-    batch_data = Tensor<T>(static_cast<size_t>(actual_batch_size),
+    batch_data = Tensor<T>(actual_batch_size,
                            cifar100_constants::NUM_CHANNELS,
                            cifar100_constants::IMAGE_HEIGHT,
                            cifar100_constants::IMAGE_WIDTH);
 
-    batch_labels = Tensor<T>(static_cast<size_t>(actual_batch_size),
+    batch_labels = Tensor<T>(actual_batch_size,
                              static_cast<size_t>(num_classes), 1, 1);
     batch_labels.fill(static_cast<T>(0.0));
 
 #pragma omp parallel for if (actual_batch_size > 16)
-    for (int i = 0; i < actual_batch_size; ++i) {
+    for (size_t i = 0; i < actual_batch_size; ++i) {
       const auto &image_data = data_[this->current_index_ + i];
 
-      for (int c = 0; c < static_cast<int>(cifar100_constants::NUM_CHANNELS);
-           ++c) {
-        for (int h = 0; h < static_cast<int>(cifar100_constants::IMAGE_HEIGHT);
-             ++h) {
-          for (int w = 0; w < static_cast<int>(cifar100_constants::IMAGE_WIDTH);
-               ++w) {
+      for (size_t c = 0; c < cifar100_constants::NUM_CHANNELS; ++c) {
+        for (size_t h = 0; h < cifar100_constants::IMAGE_HEIGHT; ++h) {
+          for (size_t w = 0; w < cifar100_constants::IMAGE_WIDTH; ++w) {
             size_t pixel_idx = c * cifar100_constants::IMAGE_HEIGHT *
                                    cifar100_constants::IMAGE_WIDTH +
                                h * cifar100_constants::IMAGE_WIDTH + w;
@@ -384,7 +381,7 @@ public:
   /**
    * Pre-compute all batches for efficient training
    */
-  void prepare_batches(int batch_size) override {
+  void prepare_batches(size_t batch_size) override {
     if (data_.empty()) {
       std::cerr << "Warning: No data loaded, cannot prepare batches!"
                 << std::endl;
